+++ Matchismo/Model/PlayingCard.m
@@ -13,7 +13,7 @@
 - (int)match:(NSArray *)otherCards
 {
     int score = 0;
-    int numOtherCards = [otherCards count];
+    NSUInteger numOtherCards = [otherCards count];
     
     if (numOtherCards) {
         for (Card *card in otherCards) {
